snd-dice: Addresses #1068
Starting all streams on first dice_pcm_hw_params call and shutting them down when all pcm devices are closed. This is to be considered as a workaround to the sync problems whencapture and playback isochronous streams are started and stopped separately. But since we're lacking DICE documentation this task would require extensive reverse engineering therefore this solution seems to be the best viable.


git-svn-id: http://192.168.0.10/svn/h10/trunk/basestation/firmware/build/alsa/sound/firewire/dice@5741 30c425cc-ffdc-49d2-aa5a-6251e47477a6