dice_sync_src_put():  don't overwrite current clock rate
When setting a different clock source, read the current clock rate first.
Before we only set the bits corresponding to the desired clock source,
putting all the clock rate related bits to zero. This was equivalent to
selecting a clock rate of 32000, which is not supported by our sound cards.